# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
...
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /launch/concierge/advice:
    post:
      operationId: createConciergeAdvice
      tags: [concierge]
      summary: Get request-only advertising format guidance
      description: Returns bounded guidance from anonymous advertising-intent signals derived in the browser. Question text, email, payment, inventory reservation, activation, and chat transcripts are not accepted or collected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConciergeAdviceInput"
      responses:
        "200":
          description: Advertising guidance or a deterministic fallback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConciergeAdviceResponse"
        "400":
          description: Invalid or oversized concierge request
...
  /launch/sponsor/reserve:
    post:
      operationId: createSponsorReservation
      tags: [launch]
      summary: Create a request-only Founding Sponsor reservation
      description: Creates a durable request for private commercial and host review. Every package is request-only and no payment is collected by this application. Clinic discount eligibility is verified privately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SponsorReservationInput"
      responses:
        "202":
          description: Reservation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReservationReceipt"
        "400":
          description: Invalid reservation request
  /launch/sponsor/status:
    get:
      operationId: getSponsorReservationStatus
      tags: [launch]
      summary: Get reservation status by private lookup token
      description: Returns the current reservation status. The token is passed as a query parameter so it is never recorded in access logs. Returns 404 for invalid or missing tokens without revealing existence.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string, minLength: 36 }
      responses:
        "200":
          description: Current reservation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SponsorReservationStatus"
        "404":
          description: Token not found
...
components:
  securitySchemes:
    ClerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk browser session for an individually authenticated staff identity.
...
  schemas:
...
    ConciergeAdviceInput:
      type: object
      additionalProperties: false
      required: [bookingIntent, awarenessIntent, comparisonIntent, privateReviewIntent, sensitiveContentDetected]
      properties:
        bookingIntent:
          type: boolean
          description: The browser detected a booking-oriented advertising question.
        awarenessIntent:
          type: boolean
          description: The browser detected an awareness-oriented advertising question.
        comparisonIntent:
          type: boolean
          description: The browser detected a request to compare advertising paths.
        privateReviewIntent:
          type: boolean
          description: The browser detected a pricing, availability, legal, contract, forecast, or outcome question that requires private review.
        sensitiveContentDetected:
          type: boolean
          description: The browser withheld possible contact, health, payment, credential, or identity information. The original text is never transmitted.
        selectedInterest:
          type: string
          enum: [performance, display, both]
          description: Optional format already being considered by the visitor.
    ConciergeAdviceResponse:
      type: object
      additionalProperties: false
      required: [message, suggestions, handoffAllowed, source]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 700
        recommendedInterest:
          type: string
          enum: [performance, display, both]
        suggestions:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: string
            enum: [performance, display, both, private_review]
        handoffAllowed:
          type: boolean
          description: True only when a validated advertising-format recommendation can preselect the request form.
        source:
          type: string
          enum: [ai, fallback]
...
    SponsorReservationInput:
      type: object
      required:
        [
          packageType,
          accessRoute,
          buyerEmail,
          buyerName,
          companyName,
          companyWebsite,
          requestedCategory,
          advertisingObjective,
          preferredPlacement,
          investmentRange,
          launchTimeline,
          creativeStatus,
          termsAccepted,
        ]
      properties:
        packageType:
          type: string
          enum: [brand, retail, other, clinic, physio, test_pilot]
          description: "brand/retail/other → Founding Sponsor (CAD 5k). clinic/physio → reserve-only, privately verified. test_pilot → CAD 10k reserve-only."
        accessRoute:
          type: string
          enum: [open_market, ecosystem_contributor, test_pilot]
          description: "Public request route. Contributor submissions remain pending manual eligibility verification; this field never grants priority by itself."
        buyerEmail:
          type: string
          minLength: 3
          maxLength: 320
        buyerName:
          type: string
          minLength: 2
          maxLength: 120
        companyName:
          type: string
          minLength: 2
          maxLength: 160
        companyWebsite:
          type: string
          minLength: 8
          maxLength: 2048
        requestedHubId:
          type: string
          minLength: 1
          maxLength: 120
        requestedHubName:
          type: string
          minLength: 1
          maxLength: 160
        requestedCategory:
          type: string
          minLength: 2
          maxLength: 80
        advertisingObjective:
          type: string
          enum: [brand_awareness, product_discovery, product_purchase, service_booking, other]
          description: Primary objective for the requested advertising campaign.
        preferredPlacement:
          type: string
          enum: [post_checkout, recovery_plan, booking_confirmation, member_hub, guidance_needed]
          description: Display moment the advertiser is most interested in.
        investmentRange:
          type: string
          enum: [under_5k, 5k_to_10k, 10k_to_25k, 25k_plus, exploring]
          description: Indicative campaign investment range for qualification only.
        launchTimeline:
          type: string
          enum: [within_30_days, 30_to_60_days, 60_plus_days, flexible]
          description: Advertiser's preferred campaign start window.
        creativeStatus:
          type: string
          enum: [creative_ready, needs_support, exploring]
          description: Whether campaign creative is ready or still being planned.
        additionalNotes:
          type: string
          maxLength: 1000
          description: Optional advertiser context for private review.
        termsAccepted:
          type: boolean
          description: Must be true. Reservation is rejected if false or absent.
        marketingConsent:
          type: boolean
          description: Optional explicit consent to receive Birch Reserve updates. Defaults to false.
...
    SponsorReservationReceipt:
      type: object
      required: [reservationId, status, lookupToken]
      properties:
        reservationId:
          type: string
          description: Opaque reservation identifier.
        status:
          type: string
          enum: [reserve_only]
        lookupToken:
          type: string
          minLength: 36
          description: Save securely. Used to poll reservation status. Not guessable.
    SponsorReservationStatus:
      type: object
      required: [reservationId, status, packageKey, amountCents, currency, createdAt]
      properties:
        reservationId:
          type: string
        status:
          type: string
          enum: [reserve_only]
        packageKey:
          type: string
        amountCents:
          type: number
        currency:
          type: string
          enum: [cad]
        createdAt:
          type: string
          format: date-time
...